fix: relax `@eslint/compat` eslint peerDependencies constraint
Updated to the suggestion in https://github.com/eslint/rewrite/pull/106#issuecomment-2525696320. This was a breaking change, causing a refresh of our lockfile to generate the following error:

```
 ERR_PNPM_PEER_DEP_ISSUES  Unmet peer dependencies

XXX
└─┬ @eslint/compat 1.2.9
  └── ✕ unmet peer eslint@^9.10.0: found 8.57.0
```

This resolves the breaking change, since I believe @eslint/compat was introduced during the v8 eslint cycle and we don't need to go lower.